NewsAnalysisBot/requirements.txt
discord.py>=2.3.0
psycopg2-binary>=2.9.0
textblob>=0.17.0
feedparser>=6.0.0
aiohttp>=3.9.0
cryptography>=41.0.0
requests>=2.31.0
